Research/backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.10.0
cymem==2.0.8
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
fastapi==0.110.0
flatbuffers==24.3.7
fonttools==4.50.0
gcloud==0.18.3
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==2.10
Jinja2==3.1.3
joblib==1.3.2
jws==0.1.3
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.3
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
openai==1.14.1
opencv-contrib-python==4.9.0.80
opencv-python==4.8.0.76
packaging==24.0
pillow==10.2.0
preshed==3.0.9
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.29.0
requests-toolbelt==0.10.1
roboflow==1.1.7
rsa==4.9
scikit-learn==1.3.2
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.4.6
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
supervision==0.16.0
thinc==8.2.3
threadpoolctl==3.3.0
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.28.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
